requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.20
sqlalchemy>=1.4
asyncpg>=0.27
databases>=0.6
pydantic>=1.10
google-genai>=0.14
langgraph>=0.1.0
slowapi>=0.1.4
structlog>=23.1.0
python-dotenv>=1.0
aiofiles
psycopg[binary]
pre-commit
black
isort
flake8